+++ tests/fixtures/fixture_recursive_strategy_list_replace.py
@@ -1,0 +1,80 @@
+
+data = [
+	(
+		{
+			"test1": 1,
+			"test2": "two",
+			"test3": True,
+			"test4": False,
+			"test5": 0.666,
+			"test6": (1, 2, 3),
+			"test7": (
+				"test",
+				42,
+				-500.5,
+				(1, 2, 3),
+				{"one": "one", "two": 2}
+			),
+			"test8": {
+				"sub_test_1": "gg-wp",
+				"sub_test_2": -12,
+				"sub_test_3": 3.1415,
+				"sub_test_4": {
+					"sub_sub_test_1": 1,
+					"sub_sub_test_2": 2,
+					"sub_sub_test_3": 3,
+					"sub_sub_test_4": (),
+					"sub_sub_test_5": {},
+					"sub_sub_test_6": "gooooo",
+				},
+			}
+		},
+
+		(
+			{
+				"test1": 4,
+				"test2": "three",
+				"test4": False,
+				"test5": 0.666,
+				"test6": (100, 200, 300),
+				"test7": (
+					42,
+					-500.5,
+					(10, 20, 30),
+					{"two": 3}
+				),
+				"test8": {
+					"sub_test_1": "gg-wp",
+					"sub_test_2": -12,
+					"sub_test_4": {
+						"sub_sub_test_1": 1,
+						"sub_sub_test_2": 2,
+						"sub_sub_test_3": 3,
+						"sub_sub_test_5": {},
+						"sub_sub_test_6": "noooooo",
+					},
+				}
+			},
+			{
+				"test1": 1,
+				"test2": "two",
+				"test3": True,
+				"test6": (1, 2, 3),
+				"test7": (
+					"test",
+					42,
+					-500.5,
+					(1, 2, 3),
+					{"one": "one", "two": 2}
+				),
+				"test8": {
+					"sub_test_3": 3.1415,
+					"sub_test_4": {
+						"sub_sub_test_4": (),
+						"sub_sub_test_6": "gooooo",
+					},
+				}
+			}
+		)
+	),
+]